Simplify maintenacnce of Bokeh release notes docs
This commit adds a :bokeh-index-toctree: that will properly sort release
docs that are included in the tocree by means of a glob. Accordingly,
the complicated logic in the deploy.py script that updated the toctree
manually has been removed.

The :bokeh-issues: directive had the format of its output upadated.

This commit also adds preliminary release notes for 0.12.4 and fixes
typos in previous release notes.